add _test.go file
