+++ precode_test.go
@@ -1,0 +1,50 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	totalCount := 4
+	req := httptest.NewRequest("GET", "/cafe?city=moscow&count=10", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	// здесь нужно добавить необходимые проверки
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+
+	cafes := strings.Split(responseRecorder.Body.String(), ",")
+	assert.Equal(t, totalCount, len(cafes))
+}
+
+func TestMainHandlerWhenEverythingOK(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?city=moscow&count=2", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+
+	assert.NotEmpty(t, responseRecorder.Body)
+}
+
+func TestMainHandlerWhenCityNotExist(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?city=ufa&count=2", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+
+	assert.Equal(t, "wrong city value", responseRecorder.Body.String())
+}
